SCRUM-165: Nuxt FE(refactor):Apply changes to account for new file system
